Finished base fee system.

+++ src/main/java/com/task/fooddelivery/dto/DeliveryFeeDto.java
@@ -1,6 +1,8 @@
 package com.task.fooddelivery.dto;
 
+import java.math.BigDecimal;
+
 public class DeliveryFeeDto {
 
-    public double fee;
+    public BigDecimal fee;
 }
